admin-dashboard/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Monitoring & Analytics
prometheus-client==0.19.0
grafana-api==1.0.3
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# AI & ML
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
tensorflow==2.14.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-dateutil==2.8.2
pydantic==2.5.2
email-validator==2.1.0.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 